requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
asgiref=3.3.1=pyhd8ed1ab_0
backports-entry-points-selectable=1.1.1=pypi_0
blinker=1.4=py_1
brotlipy=0.7.0=py38h8df0ef7_1001
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.14.4=py38ha312104_0
chardet=4.0.0=py38h578d9bd_0
coreapi=2.3.3=pypi_0
coreschema=0.0.4=pypi_0
coverage=5.5=py38h497a2fe_0
cryptography=3.3.1=py38h2b97feb_0
dataclasses=0.8=pyhc8e2a94_1
defusedxml=0.6.0=py_0
distlib=0.3.4=pypi_0
dj-rest-auth=2.1.3=pypi_0
django=3.1.3=pyhd8ed1ab_1
django-allauth=0.41.0=py_0
django-cors-headers=3.7.0=pyh44b312d_0
django-filter=2.4.0=py_0
django-rest-auth=0.9.5=py38h32f6830_0
django-templated-mail=1.1.1=pypi_0
djangorestframework=3.12.1=py_0
djangorestframework-simplejwt=4.6.0=pypi_0
djangorestframework_simplejwt=4.4.0=py_0
djoser=2.1.0=pypi_0
filelock=3.4.0=pypi_0
freetype=2.10.4=h7ca028e_0
httpie=2.1.0=py38h32f6830_1
icu=67.1=he1b5a44_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.1.1=pyhd8ed1ab_0
itypes=1.2.0=pypi_0
jinja2=2.11.3=pypi_0
jpeg=9d=h36c2ea0_0
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hed1e6ac_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgomp=9.3.0=h5dbcf3e_17
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h4f3a223_6
libuv=1.40.0=hd18ef5c_0
libwebp-base=1.1.0=h36c2ea0_3
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=pypi_0
ncurses=6.2=h58526e2_4
nodejs=15.2.1=h914e61d_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1k=h7f98852_0
pillow=8.0.1=py38h70fbd49_0
pip=20.2.4=py_0
pipenv=2021.11.23=pypi_0
platformdirs=2.4.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=py_0
pyjwt=1.7.1=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py38h924ce5b_2
python=3.8.6=h852b56e_0_cpython
python3-openid=3.1.0=py38h32f6830_1003
python_abi=3.8=1_cp38
python_http_client=3.1.0=py_0
pytz=2020.4=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
readline=8.0=he28a2e2_2
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
sendgrid=6.0.5=py_0
setuptools=49.6.0=py38h924ce5b_2
six=1.15.0=pyh9f0ad1d_0
social-auth-app-django=4.0.0=pypi_0
social-auth-core=4.1.0=pypi_0
sqlite=3.33.0=h4cf870e_1
sqlparse=0.4.1=pyh9f0ad1d_0
tk=8.6.10=hed695b0_1
uritemplate=3.0.1=pypi_0
urllib3=1.26.2=pyhd8ed1ab_0
virtualenv=20.10.0=pypi_0
virtualenv-clone=0.5.7=pypi_0
werkzeug=2.0.0=pyhd8ed1ab_0
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarn=1.22.10=0
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.5=h6597ccf_2
